% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcao_Predicao.R
\name{predizer_novos_valores}
\alias{predizer_novos_valores}
\title{Predição valores novos}
\usage{
predizer_novos_valores(nova_matriz, modelo)
}
\arguments{
\item{nova_matriz}{Matriz de novos preditores.}

\item{beta}{Coeficientes do modelo ajustado.}
}
\value{
Valores preditos para os novos dados.
}
\description{
Esta função realiza a predição de novos dados baseados no modelo ajustado.
}
