% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcao_Regressao.R
\name{regressao_linear}
\alias{regressao_linear}
\title{Regressão Linear}
\usage{
regressao_linear(x, y)
}
\arguments{
\item{x}{Matriz de preditores.}

\item{y}{Vetor de resposta.}
}
\value{
Uma lista contendo coeficientes, resíduos, valores preditos, a matriz de preditores e um gráfico.
}
\description{
Esta função realiza a regressão linear e retorna coeficientes, resíduos,
valores preditos, a matriz de preditores e um gráfico de valores preditos vs observados.
}
